Added classifiers utils, split different classes into different files
