+++ medinify/classifiers/__init__.py
@@ -1,4 +1,10 @@
 
-from .cnn_classifier import CNNLearner, ClassificationNetwork, get_lookup_table
-from .classifier import Classifier, Model
-from .classifier import print_validation_metrics
+from .cnn_learner import CNNLearner
+from .cnn_learner import ClassificationNetwork
+from .classifier import Classifier
+from .model import Model
+from .utils import print_evaluation_metrics
+from .utils import print_validation_metrics
+from .utils import find_model
+from .data_iterator import DataIterator
+
